Split stat job related C bindings in their own C files
