+++ src/unix/jbuild
@@ -174,6 +174,14 @@
     windows_bytes_read_job
     unix_write_job
     windows_write_job
+    unix_bytes_write_job
+    unix_stat_job_utils
+    unix_stat_job
+    unix_stat64_job
+    unix_lstat_job
+    unix_lstat_64_job
+    unix_fstat_job
+    unix_fstat_64_job
     not_available
   ))
   (install_c_headers (
