+++ src/unix/lwt_unix_unix.h
@@ -42,261 +42,9 @@
 #include <unistd.h>
 
 #include "unix_recv_send_utils.h"
-/* +-----------------------------------------------------------------+
-   | JOB: guess_blocking                                             |
-   +-----------------------------------------------------------------+ */
-/* +-----------------------------------------------------------------+
-   | JOB: wait_mincore                                               |
-   +-----------------------------------------------------------------+ */
-/* +-----------------------------------------------------------------+
-   | JOB: open                                                       |
-   +-----------------------------------------------------------------+ */
-/* +-----------------------------------------------------------------+
-   | JOB: read                                                       |
-   +-----------------------------------------------------------------+ */
-/* +-----------------------------------------------------------------+
-   | JOB: bytes_read                                                 |
-   +-----------------------------------------------------------------+ */
-/* +-----------------------------------------------------------------+
-   | JOB: write                                                      |
-   +-----------------------------------------------------------------+ */
+#include "unix_stat_job_utils.h"
 
 
-
-/* +-----------------------------------------------------------------+
-   | JOB: bytes_write                                                |
-   +-----------------------------------------------------------------+ */
-
-struct job_bytes_write {
-    struct lwt_unix_job job;
-    int fd;
-    char *buffer;
-    long length;
-    long result;
-    int error_code;
-};
-
-static void worker_bytes_write(struct job_bytes_write *job)
-{
-    job->result = write(job->fd, job->buffer, job->length);
-    job->error_code = errno;
-}
-
-static value result_bytes_write(struct job_bytes_write *job)
-{
-    long result = job->result;
-    LWT_UNIX_CHECK_JOB(job, result < 0, "write");
-    lwt_unix_free_job(&job->job);
-    return Val_long(result);
-}
-
-CAMLprim value lwt_unix_bytes_write_job(value val_fd, value val_buffer,
-                                        value val_offset, value val_length)
-{
-    LWT_UNIX_INIT_JOB(job, bytes_write, 0);
-    job->fd = Int_val(val_fd);
-    job->buffer = (char *)Caml_ba_data_val(val_buffer) + Long_val(val_offset);
-    job->length = Long_val(val_length);
-    return lwt_unix_alloc_job(&(job->job));
-}
-
-/* +-----------------------------------------------------------------+
-   | JOB: stat                                                       |
-   +-----------------------------------------------------------------+ */
-
-struct job_stat {
-    struct lwt_unix_job job;
-    struct stat stat;
-    int result;
-    int error_code;
-    char *name;
-    char data[];
-};
-
-static value copy_stat(int use_64, struct stat *buf)
-{
-    CAMLparam0();
-    CAMLlocal5(atime, mtime, ctime, offset, v);
-
-    atime = caml_copy_double((double)buf->st_atime +
-                             (NANOSEC(buf, a) / 1000000000.0));
-    mtime = caml_copy_double((double)buf->st_mtime +
-                             (NANOSEC(buf, m) / 1000000000.0));
-    ctime = caml_copy_double((double)buf->st_ctime +
-                             (NANOSEC(buf, c) / 1000000000.0));
-    offset = use_64 ? caml_copy_int64(buf->st_size) : Val_int(buf->st_size);
-    v = caml_alloc_small(12, 0);
-    Field(v, 0) = Val_int(buf->st_dev);
-    Field(v, 1) = Val_int(buf->st_ino);
-    switch (buf->st_mode & S_IFMT) {
-        case S_IFREG:
-            Field(v, 2) = Val_int(0);
-            break;
-        case S_IFDIR:
-            Field(v, 2) = Val_int(1);
-            break;
-        case S_IFCHR:
-            Field(v, 2) = Val_int(2);
-            break;
-        case S_IFBLK:
-            Field(v, 2) = Val_int(3);
-            break;
-        case S_IFLNK:
-            Field(v, 2) = Val_int(4);
-            break;
-        case S_IFIFO:
-            Field(v, 2) = Val_int(5);
-            break;
-        case S_IFSOCK:
-            Field(v, 2) = Val_int(6);
-            break;
-        default:
-            Field(v, 2) = Val_int(0);
-            break;
-    }
-    Field(v, 3) = Val_int(buf->st_mode & 07777);
-    Field(v, 4) = Val_int(buf->st_nlink);
-    Field(v, 5) = Val_int(buf->st_uid);
-    Field(v, 6) = Val_int(buf->st_gid);
-    Field(v, 7) = Val_int(buf->st_rdev);
-    Field(v, 8) = offset;
-    Field(v, 9) = atime;
-    Field(v, 10) = mtime;
-    Field(v, 11) = ctime;
-    CAMLreturn(v);
-}
-
-static void worker_stat(struct job_stat *job)
-{
-    job->result = stat(job->name, &job->stat);
-    job->error_code = errno;
-}
-
-static value result_stat(struct job_stat *job)
-{
-    LWT_UNIX_CHECK_JOB_ARG(job, job->result < 0, "stat", job->name);
-    value result = copy_stat(0, &job->stat);
-    lwt_unix_free_job(&job->job);
-    return result;
-}
-
-CAMLprim value lwt_unix_stat_job(value name)
-{
-    LWT_UNIX_INIT_JOB_STRING(job, stat, 0, name);
-    return lwt_unix_alloc_job(&(job->job));
-}
-
-static value result_stat_64(struct job_stat *job)
-{
-    LWT_UNIX_CHECK_JOB_ARG(job, job->result < 0, "stat", job->name);
-    value result = copy_stat(1, &job->stat);
-    lwt_unix_free_job(&job->job);
-    return result;
-}
-
-CAMLprim value lwt_unix_stat_64_job(value name)
-{
-    LWT_UNIX_INIT_JOB_STRING(job, stat, 0, name);
-    job->job.result = (lwt_unix_job_result)result_stat_64;
-    return lwt_unix_alloc_job(&(job->job));
-}
-
-/* +-----------------------------------------------------------------+
-   | JOB: lstat                                                      |
-   +-----------------------------------------------------------------+ */
-
-struct job_lstat {
-    struct lwt_unix_job job;
-    struct stat lstat;
-    int result;
-    int error_code;
-    char *name;
-    char data[];
-};
-
-static void worker_lstat(struct job_lstat *job)
-{
-    job->result = lstat(job->name, &job->lstat);
-    job->error_code = errno;
-}
-
-static value result_lstat(struct job_lstat *job)
-{
-    LWT_UNIX_CHECK_JOB_ARG(job, job->result < 0, "lstat", job->name);
-    value result = copy_stat(0, &(job->lstat));
-    lwt_unix_free_job(&job->job);
-    return result;
-}
-
-CAMLprim value lwt_unix_lstat_job(value name)
-{
-    LWT_UNIX_INIT_JOB_STRING(job, lstat, 0, name);
-    return lwt_unix_alloc_job(&(job->job));
-}
-
-static value result_lstat_64(struct job_lstat *job)
-{
-    LWT_UNIX_CHECK_JOB_ARG(job, job->result < 0, "lstat", job->name);
-    value result = copy_stat(1, &(job->lstat));
-    lwt_unix_free_job(&job->job);
-    return result;
-}
-
-CAMLprim value lwt_unix_lstat_64_job(value name)
-{
-    LWT_UNIX_INIT_JOB_STRING(job, lstat, 0, name);
-    job->job.result = (lwt_unix_job_result)result_lstat_64;
-    return lwt_unix_alloc_job(&(job->job));
-}
-
-/* +-----------------------------------------------------------------+
-   | JOB: fstat                                                      |
-   +-----------------------------------------------------------------+ */
-
-struct job_fstat {
-    struct lwt_unix_job job;
-    int fd;
-    struct stat fstat;
-    int result;
-    int error_code;
-};
-
-static void worker_fstat(struct job_fstat *job)
-{
-    job->result = fstat(job->fd, &(job->fstat));
-    job->error_code = errno;
-}
-
-static value result_fstat(struct job_fstat *job)
-{
-    LWT_UNIX_CHECK_JOB(job, job->result < 0, "fstat");
-    value result = copy_stat(0, &(job->fstat));
-    lwt_unix_free_job(&job->job);
-    return result;
-}
-
-CAMLprim value lwt_unix_fstat_job(value val_fd)
-{
-    LWT_UNIX_INIT_JOB(job, fstat, 0);
-    job->fd = Int_val(val_fd);
-    return lwt_unix_alloc_job(&(job->job));
-}
-
-static value result_fstat_64(struct job_fstat *job)
-{
-    LWT_UNIX_CHECK_JOB(job, job->result < 0, "fstat");
-    value result = copy_stat(1, &(job->fstat));
-    lwt_unix_free_job(&job->job);
-    return result;
-}
-
-CAMLprim value lwt_unix_fstat_64_job(value val_fd)
-{
-    LWT_UNIX_INIT_JOB(job, fstat, 0);
-    job->job.result = (lwt_unix_job_result)result_fstat_64;
-    job->fd = Int_val(val_fd);
-    return lwt_unix_alloc_job(&(job->job));
-}
 
 /* +-----------------------------------------------------------------+
    | JOB: utimes                                                     |
