+++ src/unix/windows_c/not_available.c
@@ -59,4 +59,10 @@ LWT_NOT_AVAILABLE2(unix_set_affinity)
 LWT_NOT_AVAILABLE1(unix_guess_blocking_job)
 LWT_NOT_AVAILABLE2(unix_wait_mincore_job)
 LWT_NOT_AVAILABLE1(unix_open_job)
+LWT_NOT_AVAILABLE1(unix_stat_job)
+LWT_NOT_AVAILABLE1(unix_stat_64_job)
+LWT_NOT_AVAILABLE1(unix_lstat_job)
+LWT_NOT_AVAILABLE1(unix_lstat_64_job)
+LWT_NOT_AVAILABLE1(unix_fstat_job)
+LWT_NOT_AVAILABLE1(unix_fstat_64_job)
 #endif
